Quicksand
Se creo script de quicksand mas simple, se agregaron animaciones mas action para la muerte del pj
