+++ Assets/Prefabs/Modules/Lvl1Quicksand.prefab
@@ -12,8 +12,8 @@ GameObject:
   - component: {fileID: 5143225568058333746}
   - component: {fileID: 6360575976088458270}
   - component: {fileID: 4703367273491461837}
-  - component: {fileID: 5972239853972133095}
   - component: {fileID: 7706660337637947504}
+  - component: {fileID: 786889945452448846}
   m_Layer: 0
   m_Name: Lvl1Quicksand
   m_TagString: Untagged
@@ -29,7 +29,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2801768664947367953}
   serializedVersion: 2
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
@@ -100,7 +100,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   m_PolyMesh:
-    name: PolybrushMesh34372
+    name: PolybrushMesh-264930
     vertices:
     - {x: 0.9185412, y: 1.1365021, z: 12.137853}
     - {x: -0.9185427, y: -14.539595, z: 12.137853}
@@ -486,7 +486,7 @@ MonoBehaviour:
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
-    - {r: 0.8117647, g: 0.8117647, b: 0.8117647, a: 1}
+    - {r: 0.81200856, g: 0.81200856, b: 0.81200856, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
@@ -507,6 +507,7 @@ MonoBehaviour:
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
+    - {r: 0.8271413, g: 0.8271413, b: 0.8271413, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
@@ -554,6 +555,7 @@ MonoBehaviour:
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
+    - {r: 0.8271413, g: 0.8271413, b: 0.8271413, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
@@ -564,24 +566,22 @@ MonoBehaviour:
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
+    - {r: 0.11337541, g: 0.11337541, b: 0.11337541, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
-    - {r: 0.11372549, g: 0.11372549, b: 0.11372549, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
+    - {r: 0.000014650606, g: 0.000014650606, b: 0.000014650606, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
+    - {r: 0.81200856, g: 0.81200856, b: 0.81200856, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
-    - {r: 0, g: 0, b: 0, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
-    - {r: 0.8117647, g: 0.8117647, b: 0.8117647, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
-    - {r: 1, g: 1, b: 1, a: 1}
-    - {r: 1, g: 1, b: 1, a: 1}
-    - {r: 1, g: 1, b: 1, a: 1}
+    - {r: 0.64712596, g: 0.64712596, b: 0.64712596, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
     - {r: 1, g: 1, b: 1, a: 1}
@@ -1027,23 +1027,6 @@ MonoBehaviour:
   m_SkinMeshRef: {fileID: 0}
   m_OriginalMeshObject: {fileID: 3919423610170832201, guid: b8ae31481fb552e45a91cb59f4d65f86, type: 3}
   m_Mode: 0
---- !u!114 &5972239853972133095
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2801768664947367953}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2f4c0fa648712e34ba20cf7e7c34cf54, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  moveSpeedDownNormal: 0.2
-  moveSpeedDownSmall: 0.1
-  moveSpeedDownHead: 0
-  moveSpeedUp: 0.5
-  yOffset: -5
 --- !u!65 &7706660337637947504
 BoxCollider:
   m_ObjectHideFlags: 0
@@ -1051,7 +1034,7 @@ BoxCollider:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2801768664947367953}
-  m_Material: {fileID: 0}
+  m_Material: {fileID: 13400000, guid: f2cc16ad81f6028489433eacceb90bdb, type: 2}
   m_IncludeLayers:
     serializedVersion: 2
     m_Bits: 0
@@ -1065,6 +1048,21 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 13.3672695, y: 15.676097, z: 24.275705}
   m_Center: {x: 0.25384188, y: -6.701546, z: 0}
+--- !u!114 &786889945452448846
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2801768664947367953}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 80f7251162fcf764580bc10dc29746e4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _timeDeathNormal: 2
+  _timeDeathSmall: 2
+  _sinkFX: {fileID: 0}
 --- !u!1 &3307113323285950692
 GameObject:
   m_ObjectHideFlags: 0
